add error handling for Greycite API

+++ R/greycite.R
@@ -10,11 +10,16 @@
 greycite <- function(url, format=c("bib","json")){
   format <- match.arg(format)
   dat <- GET(paste("http://greycite.knowledgeblog.org/", format, "/?uri=", url, sep=""))
-  if(format == "bib"){
-    con <- tempfile("greycite.bib")
-    writeLines(content(dat, "text"), con=con)
-    bib <- ReadBib(con)
-    unlink(con)
+  if(dat$status_code == 200){
+    if(format == "bib"){
+      con <- tempfile("greycite.bib")
+      writeLines(content(dat, "text"), con=con)
+      bib <- ReadBib(con)
+      unlink(con)
+    }
+  } else {
+    warning("cannot reach Greycite API")
+    bib <- bibentry(bibtype = "misc", author="Unknown", title="Unknown", url = url, howpublished=paste0("\\url{", url, "}"), year=format(Sys.Date(), "%Y"), note = paste("Accessed", Sys.Date()))
   }
   bib
 }
